moved eventbus and handler to common
