+++ src/common/domain/events/EventType.ts
@@ -1,0 +1,7 @@
+/**
+ * Value Object interface for domain event types.
+ * Each bounded context defines its own EventType implementation.
+ */
+export interface EventType {
+    readonly value: string;
+}
