+++ src/common/usecase/EventBus.ts
@@ -1,8 +1,8 @@
 import { DomainEvent } from '#common/domain/events/DomainEvent';
-import { LobbyEventType } from '#matchmaking/domain/lobby/events/LobbyEventType';
+import { EventType } from '#common/domain/events/EventType';
 import { EventHandler } from './EventHandler';
 
 export interface EventBus {
     publish<T extends DomainEvent>(event: T): void;
-    register<T extends DomainEvent>(eventType: LobbyEventType, handler: EventHandler<T>): void;
+    register<T extends DomainEvent>(eventType: EventType, handler: EventHandler<T>): void;
 }
