+++ src/matchmaking/infastructure/handlers/WsNotifyPlayerLeftLobbyHandler.ts
@@ -1,5 +1,5 @@
+import { EventHandler } from '#common/usecase/EventHandler';
 import { PlayerLeftLobby } from '#matchmaking/domain/lobby/events/PlayerLeftLobby';
-import { EventHandler } from '../../usecase/EventHandler';
 
 export class WsNotifyPlayerLeftLobbyHandler implements EventHandler<PlayerLeftLobby> {
     handle(event: PlayerLeftLobby): void {
